requirements.txt
pytest==7.4.0
selenium==4.18.1
openpyxl==3.1.2
allure-pytest==2.13.2
allure-python-commons==2.13.2
openpyxl==3.1.2
pytest==7.4.0
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-xdist==3.3.1
selenium==4.18.1
